+++ CHANGELOG.md
@@ -1,3 +1,17 @@
 ## 0.0.1
 
 Added one animation (side-way surge).
+
+## 0.0.2
+
+Update readme
+
+## 0.0.3
+
+Fix Sideway Surge animation
+
+## 0.0.4
+
+Introduced Circle Surge animation
+Introduced Dual Surge animation
+Introduced Face Surge animation